+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
         tools:targetApi="31">
         <activity
             android:screenOrientation="portrait"
-            android:name=".MainActivity"
+            android:name=".ui.activities.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.Hiker">
             <intent-filter>
