+++ app/src/main/java/com/example/hiker/utils/DistanceCalculator.kt
@@ -1,0 +1,19 @@
+package com.example.hiker.utils
+
+object DistanceCalculator {
+    fun calculateDistance(lat1: Double, lon1: Double, lat2: Double, lon2: Double): Float {
+        val earthRadius = 6371000.0 // Rayon de la Terre en mètres
+        val lat1Rad = Math.toRadians(lat1)
+        val lon1Rad = Math.toRadians(lon1)
+        val lat2Rad = Math.toRadians(lat2)
+        val lon2Rad = Math.toRadians(lon2)
+
+        val dLon = lon2Rad - lon1Rad
+        val dLat = lat2Rad - lat1Rad
+        val a = kotlin.math.sin(dLat / 2) * kotlin.math.sin(dLat / 2) +
+                kotlin.math.cos(lat1Rad) * kotlin.math.cos(lat2Rad) *
+                kotlin.math.sin(dLon / 2) * kotlin.math.sin(dLon / 2)
+        val c = 2 * kotlin.math.atan2(kotlin.math.sqrt(a), kotlin.math.sqrt(1 - a))
+        return (earthRadius * c).toFloat()
+    }
+}
